CEQ Manual for writing Do Files, Sean Higgins
Version April 2017
Introduction
This manual is intended to assist the construction of an Stata dole for a CEQ analysis. Below are the sections
that tend to fall in a CEQ do le by order of which they appear. Note that not every section falls in each do le as
some depend on the specic function of the do le. This manual was based on the set of do les that contributed
to the project "Taxes and Social Spending in Brazil" but can be followed to construct any CEQ do le.